requirements.txt
python-dotenv
openai
anthropic
pydantic
pytest
